v.6: split into simpler components
v.6: split into simpler components: FormEditor to Form Editor and Form Element
