+++ src/index.css
@@ -7,7 +7,6 @@ body {
 
 #root {
 	height: 100vh;
-	/* width: 100vw; */
 	display: flex;
 	flex-direction: column;
 	width: 98vw;
